Exp2: easy creation of practise and alpha version

+++ utils/files_delete.py
@@ -11,6 +11,7 @@ This script will delete files from a directory.
 If you instead want to move files into another directory, use files_move.py
 If you instead want to copy files into another directory, use files_copy.py
 
+The script will ask you to confirm to delete files.
 
 """
 
@@ -28,11 +29,11 @@ thisDir = os.getcwd()
 baseDir = os.path.join(thisDir[:thisDir.find('Scripts')], 'Stimuli','AudioGens','Experiment2', 'tts-golang-44100hz')
 
 dirinput = os.path.join(baseDir,'tts-golang-textGrid')
-diroutput = os.path.join(baseDir, '_delete', str(datetime.now())[:-10].replace(':',''))
+backupDir = os.path.join(baseDir, '_delete', str(datetime.now())[:-10].replace(':',''))
 
 # make directory if it doesn't exist yet
-if not os.path.isdir(diroutput):
-    os.makedirs(diroutput)
+if not os.path.isdir(backupDir):
+    os.makedirs(backupDir)
 
 #%% find target files
 files = [files for files in glob(os.path.join(dirinput,'*'), recursive=True) if re.findall('Ham|Sch|Tel|Loe',files)]
@@ -40,19 +41,20 @@ files = [files for files in glob(os.path.join(dirinput,'*'), recursive=True) if 
 
 
 #%%
+msg = '--> are you sure you want to delete ' + str(len(files)) + ' files? [y/n]'
 
-
-for file in files:
+if input(msg) == 'y':      
+    for file in files:
+        if create_backup:
+            file_name = os.path.basename(file)
+            destination = os.path.join(backupDir, file_name)
+            shutil.move(file, destination)
+        else:
+            os.remove(file)
+        print(file)
+        
     if create_backup:
-        file_name = os.path.basename(file)
-        destination = os.path.join(diroutput, file_name)
-        shutil.move(file, destination)
+        print('above files have been moved to ', backupDir)
+        print('to fully delete files, delete the above directory.')
     else:
-        os.remove(file)
-    print(file)
-    
-if create_backup:
-    print('above files have been moved to ', diroutput)
-    print('to fully delete files, delete the above directory.')
-else:
-    print('above files have been deleted')
+        print('above files have been deleted')
